Created Bootcamp and User relationship. Added Bootcamp, Courses ownership functionality to update and delete. Added forgot password & generate token and send mail using nodemailer and mailtrap.io. Added reset password and update user details functionality. Created Admin functionaltiy to perfrom CRUD operations on Users.

+++ controllers/bootcamps.js
@@ -32,6 +32,19 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
   
+    //Add user to req.body
+    req.body.user = req.user.id;
+
+    //check for published bootcamp -> 1 user should publish 1 bootcamo 
+    const publishedBootcamp = await Bootcamp.findOne({user : req.user.id});
+
+
+    //If user is not an admin then only can add 1 boocamp
+
+    if(publishedBootcamp && req.user.role !=='admin'){
+      return next(new ErrorResponse(`The user with id ${req.user.id} has already published a bootcamp`));
+    }
+
     const bootcamp = await Bootcamp.create(req.body);
 
     res.status(201).json({
@@ -47,15 +60,23 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 exports.updateBootcamp = asyncHandler(async(req, res, next) => {
   
   
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id,req.body, {
-      new : true,
-      runValidators : true
-    });
+    let bootcamp = await Bootcamp.findById(req.params.id);
   
   
     if(!bootcamp){
       return next(new ErrorResponse(`Bootcamp not fount with id of ${req.params.id}`, 404));
     }
+
+    //Make sure user is Bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+      return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp}`, 404));
+    }
+
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id ,req.body, {
+      new : true,
+      runValidators : true
+    });
   
     res.status(200).json({success : true, data : bootcamp});
  
@@ -77,6 +98,13 @@ exports.deleteBootcamp = asyncHandler( async (req, res, next) => {
     if(!bootcamp){
       return next(new ErrorResponse(`Bootcamp not fount with id of ${req.params.id}`, 404));
     }
+
+
+    //Make sure user is Bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+      return next(new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp}`, 404));
+    }
+
 
     bootcamp.deleteOne();
   
@@ -100,6 +128,10 @@ exports.bootcampPhotoUpload = asyncHandler( async (req, res, next) => {
     return next(new ErrorResponse(`Bootcamp not fount with id of ${req.params.id}`, 404));
   }
 
+  //Make sure user is Bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(new ErrorResponse(`User ${req.params.id} is not authorized to updload this bootcamp photo}`, 404));
+  }
   
   if(!req.files)
   {
