+++ controllers/courses.js
@@ -66,12 +66,18 @@ exports.getCourse = asyncHandler(async(req,res,next) => {
 exports.addCourse = asyncHandler(async(req,res,next) => {
 
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if(!bootcamp){
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404);
+    }
+
+    //Make sure user is Bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}}`, 404));
     }
 
     const course = await Course.create(req.body);
@@ -96,6 +102,12 @@ exports.updateCourse = asyncHandler(async(req,res,next) => {
     if(!course){
         return next(new ErrorResponse(`No Course with the id of ${req.params.id}`), 404);
     }
+
+
+    //Make sure user is Bootcamp owner
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}}`, 404));
+        }
 
     
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -124,6 +136,10 @@ exports.deleteCourse = asyncHandler(async(req,res,next) => {
         return next(new ErrorResponse(`No Course with the id of ${req.params.id}`), 404);
     }
 
+     //Make sure user is Bootcamp owner
+     if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}}`, 404));
+        }
     
     // instead of remove
     await course.deleteOne();
